multi: separate fees out into own package and move special case check
We will need our fee calculation logic for other subsystems, so we
pull this functionality out into its own package. The special case of
not supporting batched transactions is moved out of our fee calculation
logic and into the resolutions package where it applies.
